Little change #2

+++ E-Store2021/ViewModels/ShoppingCartModel.cs
@@ -4,7 +4,7 @@ namespace E_Store2021.Models
 {
     public static class ShoppingCartModel
     {
-        public static double? Total { get; set; }
+        public static decimal? Total { get; set; }
 
         public static int? Count { get; set; }
         public static ICollection<ShoppingCartItem> ShoppingCartItems { get; set; }
